requirements.txt
Flask==3.0.0
torch==2.0.0
transformers==4.30.0
Pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3
Werkzeug==3.0.1
Jinja2==3.1.2
uuid==1.30
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.3
setuptools==67.8.0
